refactor: split sample source db and scanner
