+++ src/sample_sources/scanner/metadata.rs
@@ -1,0 +1,80 @@
+use std::{
+    fs,
+    io::Read,
+    path::{Path, PathBuf},
+    time::{SystemTime, UNIX_EPOCH},
+};
+
+use super::scan::ScanError;
+
+#[derive(Debug)]
+pub(super) struct FileFacts {
+    pub(super) relative: PathBuf,
+    pub(super) size: u64,
+    pub(super) modified_ns: i64,
+}
+
+pub(super) fn read_facts(root: &Path, path: &Path) -> Result<FileFacts, ScanError> {
+    let relative = strip_relative(root, path)?;
+    let meta = path.metadata().map_err(|source| ScanError::Io {
+        path: path.to_path_buf(),
+        source,
+    })?;
+    let modified_ns = to_nanos(
+        &meta.modified().map_err(|source| ScanError::Io {
+            path: path.to_path_buf(),
+            source,
+        })?,
+        path,
+    )?;
+    Ok(FileFacts {
+        relative,
+        size: meta.len(),
+        modified_ns,
+    })
+}
+
+pub(super) fn compute_content_hash(path: &Path) -> Result<String, ScanError> {
+    let mut file = fs::File::open(path).map_err(|source| ScanError::Io {
+        path: path.to_path_buf(),
+        source,
+    })?;
+    let mut hasher = blake3::Hasher::new();
+    let mut buffer = [0u8; 64 * 1024];
+    loop {
+        let read = file.read(&mut buffer).map_err(|source| ScanError::Io {
+            path: path.to_path_buf(),
+            source,
+        })?;
+        if read == 0 {
+            break;
+        }
+        hasher.update(&buffer[..read]);
+    }
+    Ok(hasher.finalize().to_hex().to_string())
+}
+
+pub(super) fn is_supported_audio(path: &Path) -> bool {
+    let Some(ext) = path.extension().and_then(|ext| ext.to_str()) else {
+        return false;
+    };
+    match ext.to_ascii_lowercase().as_str() {
+        "wav" | "aif" | "aiff" | "flac" | "mp3" => true,
+        _ => false,
+    }
+}
+
+fn strip_relative(root: &Path, path: &Path) -> Result<PathBuf, ScanError> {
+    path.strip_prefix(root)
+        .map(PathBuf::from)
+        .map_err(|_| ScanError::InvalidRoot(path.to_path_buf()))
+}
+
+fn to_nanos(time: &SystemTime, path: &Path) -> Result<i64, ScanError> {
+    let duration = time
+        .duration_since(UNIX_EPOCH)
+        .map_err(|_| ScanError::Time {
+            path: path.to_path_buf(),
+        })?;
+    Ok(duration.as_nanos().min(i64::MAX as u128) as i64)
+}
